+++ src/java/utility/email_factory/CheckInBuilder.java
@@ -1,5 +1,6 @@
 package utility.email_factory;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -18,11 +19,11 @@ public class CheckInBuilder implements EmailContentBuilder{
         String endDate = (String) data.get("endDate");
         List<String> typeRoom = (List<String>) data.get("typeRoom");
         List<Integer> quantityTypeRoom = (List<Integer>) data.get("quantityTypeRoom");
-        List<Double> priceTypeRoom = (List<Double>) data.get("priceTypeRoom");
+        List<BigInteger> priceTypeRoom = (List<BigInteger>) data.get("priceTypeRoom");
         List<String> services = (List<String>) data.get("services");
         List<Integer> serviceQuantity = (List<Integer>) data.get("serviceQuantity");
-        List<Integer> servicePrice = (List<Integer>) data.get("servicePrice");
-        int total = (int) data.get("total");
+        List<BigInteger> servicePrice = (List<BigInteger>) data.get("servicePrice");
+        BigInteger total = (BigInteger) data.get("total");
         String paymentMethod = (String) data.get("paymentMethod");
         StringBuilder confirmCheckin = new StringBuilder();
         // Tính số đêm giữa startDate và endDate (định dạng yyyy-MM-dd)
